Avoid processing hbs loader twice for same file
A loader declaration with test /.hbs/ is needed for
hbs partials to be processed but the root templates
are required currently with the 'hbs!' loader syntax.

In this situation, the root hbs templates were being processed
twice, once with the `hbs!` loader syntax and once with the loader
declaration in the `modules.rules` section of the webpack config.

The `normalExecuted` attribute of each loader declaration for this file
is used to determine whether this is the first or second run of this
particular loader.